requirements.txt
streamlit
gspread
google-genai
google-oauth2
toml
requests
beautifulsoup4
